refactor: Build Error 수정 #7
- device_info 모듈 패키지 네임 수정
- settings.grdles core 모듈 제거
- FeedHttp provider module 선언
