+++ build-logic/src/main/kotlin/convention/PresentationConvention.kt
@@ -52,6 +52,7 @@ class PresentationConvention : Plugin<Project> {
             add("implementation", libs.findLibrary("coil-network").get())
             add("implementation", libs.findLibrary("hilt-android").get())
             add("implementation", libs.findLibrary("lottie-compose").get())
+            add("implementation", libs.findLibrary("hilt-navigation-compose").get())
             add("ksp", libs.findLibrary("hilt-compiler").get())
             add("debugImplementation", libs.findLibrary("androidx-ui-tooling").get())
             add("debugImplementation", libs.findLibrary("androidx-ui-test-manifest").get())
