+++ presentation/home/build.gradle.kts
@@ -9,4 +9,5 @@ android {
 dependencies {
     implementation(libs.paging.compose)
     implementation(libs.paging.runtime)
+    implementation(libs.hilt.navigation.compose)
 }
