+++ settings.gradle.kts
@@ -25,7 +25,6 @@ include(":app")
 include(":presentation")
 include(":domain")
 include(":data")
-include(":core")
 include(":data:network")
 include(":data:device")
 include(":core:core-design")
